LISTA 2 ATIVIDADE 4 MODIFICADA

+++ java/AULAS/src/lista2/Atividade4.java
@@ -10,9 +10,9 @@ public class Atividade4 {
 		double numero;
 		
 		System.out.printf("Digite um valor inteiro positivo: ");
-		numero = scan.nextInt();
+		numero = scan.nextDouble();
 		
-		if (numero< 0) {
+		if (numero<0) {
 			System.out.printf("\nImpossível...número negativo.");
 		}
 		else if (numero==0) {
